pyproject.toml
[project]
name = "genie-audit-streamlit"
version = "1.0.0"
description = "Genie Performance Audit - Streamlit app for analyzing Databricks Genie query performance"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Databricks", email = "support@databricks.com" }
]

dependencies = [
    "streamlit>=1.41.0",
    "databricks-sdk>=0.40.0",
    "plotly>=5.24.0",
    "pandas>=2.2.0",
    "fpdf2>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "integration: mark test as an integration test (requires Databricks credentials)",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
log_cli = false
log_cli_level = "INFO"

[tool.hatch.build.targets.wheel]
packages = ["services", "utils", "components"]
